reroute-ai/backend/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "reroute-backend"
version = "0.1.0"
description = "ReRoute.AI FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "bcrypt>=4.0.0",
    "PyJWT>=2.8.0",
    "httpx>=0.27.0",
    "email-validator>=2.0.0",
]

[project.optional-dependencies]
agent = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
]

[tool.setuptools]
py-modules = ["main", "config", "api_router", "deps", "database"]

[tool.setuptools.packages.find]
include = ["routers", "service", "dao", "schema", "model", "integrations", "agent", "utils"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["."]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
